refactor: refactor history, favorites and card pages
