+++ src/pages/history/history.tsx
@@ -1,34 +1,9 @@
 import { HistoryItem, Loader } from '@components/index';
-import { useAppDispatch } from '@hooks/index';
-import { AppRoutes } from '@router/routes';
-import { fetchHistory } from '@store/historySlice';
-import { selectHistory, selectUseIsLoggedIn } from '@store/selectors';
-import { useEffect } from 'react';
-import { useSelector } from 'react-redux';
-import { useNavigate } from 'react-router-dom';
+import { useHistory } from '@hooks/useHistory';
 import s from './history.module.scss';
 
 export const HistoryPage = () => {
-  const isLoggedIn = useSelector(selectUseIsLoggedIn);
-  const history = useSelector(selectHistory);
-  const navigate = useNavigate();
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    const getHistory = async () => {
-      await dispatch(fetchHistory());
-    };
-
-    if (isLoggedIn) {
-      getHistory();
-    }
-  }, [dispatch, isLoggedIn]);
-
-  useEffect(() => {
-    if (isLoggedIn === false) {
-      navigate(AppRoutes.LOGIN_ROUTE);
-    }
-  }, [isLoggedIn, navigate]);
+  const { history } = useHistory();
 
   if (!history) {
     return <Loader />;
